Support attribute content on all fields.
<sometag content="something" /> should be considered, regardless if `content` is not a valid
attribute of `sometag`.

The specification for microdata[1] details that an elements content attribute should be considered
before text content.

Any23 doesn't currently do this, it only considers `content` for `meta` tags which is the only
HTML tag which is suppose to have a `content` but not all sites follow HTML specifications.

Updating the microdata parser to be able to get `content` from any element should it exist.

[1] https://www.w3.org/TR/microdata/#values

Signed-off-by: Ian Duffy <ian.duffy@zalando.ie>